Factored out LifecycleHandler baseclass
