+++ bokeh/application/handlers/server_lifecycle.py
@@ -31,7 +31,7 @@ import os
 # Bokeh imports
 from ...util.callback_manager import _check_callback
 from .code_runner import CodeRunner
-from .handler import Handler
+from .lifecycle import LifecycleHandler
 
 #-----------------------------------------------------------------------------
 # Globals and constants
@@ -41,7 +41,7 @@ from .handler import Handler
 # General API
 #-----------------------------------------------------------------------------
 
-class ServerLifecycleHandler(Handler):
+class ServerLifecycleHandler(LifecycleHandler):
     ''' Load a script which contains server lifecycle callbacks.
 
     '''
@@ -66,11 +66,6 @@ class ServerLifecycleHandler(Handler):
         source = codecs.open(filename, 'r', 'UTF-8').read()
 
         self._runner = CodeRunner(source, filename, argv)
-
-        self._on_server_loaded = _do_nothing
-        self._on_server_unloaded = _do_nothing
-        self._on_session_created = _do_nothing
-        self._on_session_destroyed = _do_nothing
 
         if not self._runner.failed:
             # unlike ScriptHandler, we only load the module one time
@@ -119,66 +114,6 @@ class ServerLifecycleHandler(Handler):
 
     # Public methods ----------------------------------------------------------
 
-    def modify_document(self, doc):
-        ''' This handler does not make any modifications to the Document.
-
-        Args:
-            doc (Document) : A Bokeh Document to update in-place
-
-                *This handler does not modify the document*
-
-        Returns:
-            None
-
-        '''
-        # we could support a modify_document function, might be weird though.
-        pass
-
-    def on_server_loaded(self, server_context):
-        ''' Execute `on_server_unloaded`` from the configured module (if
-        it is defined) when the server is first started.
-
-        Args:
-            server_context (ServerContext) :
-
-        '''
-        return self._on_server_loaded(server_context)
-
-    def on_server_unloaded(self, server_context):
-        ''' Execute ``on_server_unloaded`` from the configured module (if
-        it is defined) when the server cleanly exits. (Before stopping the
-        server's ``IOLoop``.)
-
-        Args:
-            server_context (ServerContext) :
-
-        .. warning::
-            In practice this code may not run, since servers are often killed
-            by a signal.
-
-        '''
-        return self._on_server_unloaded(server_context)
-
-    def on_session_created(self, session_context):
-        ''' Execute ``on_session_created`` from the configured module (if
-        it is defined) when a new session is created.
-
-        Args:
-            session_context (SessionContext) :
-
-        '''
-        return self._on_session_created(session_context)
-
-    def on_session_destroyed(self, session_context):
-        ''' Execute ``on_session_destroyed`` from the configured module (if
-        it is defined) when a new session is destroyed.
-
-        Args:
-            session_context (SessionContext) :
-
-        '''
-        return self._on_session_destroyed(session_context)
-
     def url_path(self):
         ''' The last path component for the basename of the path to the
         callback module.
@@ -197,9 +132,6 @@ class ServerLifecycleHandler(Handler):
 #-----------------------------------------------------------------------------
 # Private API
 #-----------------------------------------------------------------------------
-
-def _do_nothing(ignored):
-    pass
 
 #-----------------------------------------------------------------------------
 # Code
